+++ CHANGELOG.md
@@ -42,6 +42,8 @@
 
 - fix `PmxLoader`, `PmdLoader` crash when load model with no meshes
 
+- fix crash when evaluating texture alpha on model with no submeshes
+
 ## 0.68.0 (2025-07-16)
 
 - BPMX format updated to 3.0.0. 2.X files are no longer supported